добавлено cancelEvent

+++ main-server/src/main/java/ru/practicum/app/event/EventRepository.java
@@ -2,7 +2,6 @@ package ru.practicum.app.event;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.practicum.app.event.model.Event;
-import ru.practicum.app.user.User;
 
-public interface EventRepository extends JpaRepository<Event, Long> {
+public interface EventRepository extends JpaRepository<Event, Integer> {
 }
